Split out the see more button
