+++ react/components/Autocomplete/components/TileList/TileList.tsx
@@ -1,11 +1,11 @@
 import React, { FC } from 'react'
-import { ExtensionPoint, Link } from 'vtex.render-runtime'
 import { Spinner } from 'vtex.styleguide'
 import ProductSummary from 'vtex.product-summary/ProductSummaryCustom'
 
 import styles from './styles.css'
 import CustomListItem from '../CustomListItem/CustomListItem'
 import { ProductLayout } from '../..'
+import SeeMoreButton from '../SeeMoreButton'
 
 interface TileListProps {
   term: string
@@ -24,46 +24,12 @@ interface TileListProps {
   }>
 }
 
-interface SeeMoreButtonProps {
-  term: string
-  onSeeAllClick: (term: string) => void
-  totalProducts: number
-}
-const SeeMoreButton = (props: SeeMoreButtonProps) => {
-  const { term, totalProducts } = props
-
-  // TODO Remove hard coding
-
-  const stylesObj = {
-    display: 'block',
-    height: '60px',
-    lineHeight: '60px',
-    fontSize: '13px',
-    color: '#787878',
-    textAlign: 'center',
-  }
-  return (
-    <Link
-      query={`map=ft&_q=${props.term}`}
-      params={{
-        term,
-      }}
-      page="store.search"
-      onClick={() => props.onSeeAllClick(term)}
-      styles={stylesObj}
-    >
-      See all {totalProducts} productszzzzz
-    </Link>
-  )
-}
-
 const TileList: FC<TileListProps> = ({
   term,
   title,
   products,
   showTitle,
   totalProducts,
-  layout,
   isLoading,
   onProductClick,
   onSeeAllClick,
@@ -87,8 +53,7 @@ const TileList: FC<TileListProps> = ({
           <ul
             className={styles.tileListList}
             style={{
-              flexDirection:
-                layout === ProductLayout.Horizontal ? 'column' : 'row',
+              flexDirection: 'column',
             }}
           >
             {products.map((product, index: number) => {
@@ -98,27 +63,17 @@ const TileList: FC<TileListProps> = ({
 
               return (
                 <li key={product.productId} className={styles.tileListItem}>
-                  {layout === ProductLayout.Horizontal ? (
-                    HorizontalProductSummary ? (
-                      <HorizontalProductSummary
-                        product={productSummary}
-                        actionOnClick={() => {
-                          onProductClick(productSummary.productId, index)
-                        }}
-                      />
-                    ) : (
-                      <CustomListItem
-                        product={productSummary}
-                        onClick={() => {
-                          onProductClick(productSummary.productId, index)
-                        }}
-                      />
-                    )
-                  ) : (
-                    <ExtensionPoint
-                      id="product-summary"
+                  {HorizontalProductSummary ? (
+                    <HorizontalProductSummary
                       product={productSummary}
                       actionOnClick={() => {
+                        onProductClick(productSummary.productId, index)
+                      }}
+                    />
+                  ) : (
+                    <CustomListItem
+                      product={productSummary}
+                      onClick={() => {
                         onProductClick(productSummary.productId, index)
                       }}
                     />
